Refactor organization of api functions and tests
