+++ tests/unit/test_station.py
@@ -33,36 +33,3 @@ def test_check_init_stations(app, clear_db, mock_get_ubike_json):
 
             station_rows = Station.query.all()
             assert len(station_rows) == 3
-
-
-def test_routes_index(client, mock_get_ubike_json):
-    response = client.get('/station/')
-    assert response.is_json
-    assert 'retVal' in response.get_json()
-
-
-def test_routes_get_by_station_number(app, client, mock_get_ubike_json):
-    with app.app_context():
-        station.check_init_stations()
-
-    response_exist = client.get('/station/1')
-    assert response_exist.is_json
-    assert 'sno' in response_exist.get_json()
-
-    response_not_exist = client.get('/station/4')
-    assert response_not_exist.status_code == 404
-
-    # TODO: 404 condition
-    # TODO: return json if error
-
-
-def test_routes_get_by_station_name(app, client, mock_get_ubike_json):
-    with app.app_context():
-        station.check_init_stations()
-
-    response_exist = client.get('/station/捷運市政府站(3號出口)')
-    assert response_exist.is_json
-    assert 'sno' in response_exist.get_json()
-
-    response_not_exist = client.get('/station/abc')
-    assert response_not_exist.status_code == 404
